update README.md file and solve error on webpack build
