--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "to-do-list-project",
-  "version": "1.0.0",
-  "description": "![](https://img.shields.io/badge/Microverse-blueviolet)",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/PNdunguMaina/To-do-list-project.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/PNdunguMaina/To-do-list-project/issues"
-  },
-  "homepage": "https://github.com/PNdunguMaina/To-do-list-project#readme",
-  "devDependencies": {
-    "eslint": "^8.20.0"
-  }
-}
