+++ webpack/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "boilerplate",
+  "name": "webpack-demo",
   "version": "1.0.0",
   "description": "",
   "private": true,
