+++ webpack/src/index.html
@@ -3,10 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css"
-    />
+    
     <script
       src="https://kit.fontawesome.com/03046ec777.js"
       crossorigin="anonymous"
@@ -33,6 +30,5 @@
       </form>
       
     </div>
-    <script type="module" src="./modules/event.js"></script>
   </body>
 </html>
