--- webpack/src/modules/main.js
@@ -1,8 +1,0 @@
-// Template class object
-export default class ToDoTasks {
-    constructor(description, completed, index) {
-      this.description = description;
-      this.completed = completed;
-      this.index = index;
-    }
-  }
