+++ webpack/src/modules/mainClassContainer.js
@@ -1,0 +1,8 @@
+// Template class object
+export default class ToDoTasks {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
+  }
+}
